Improved saving/loading NBT
I have not really solved the problem completely. But it's working a lot better now, and it's down to logical errors in my code. The load/save functionality is actually being called. I probably want to yeet some of the code in TechnicalManager.kt that uses the "previous" save info (I had built it that way to be robust to schema changes in the future)